% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty.R
\name{emptyArray}
\alias{emptyArray}
\title{emptyArray}
\usage{
emptyArray(dim = length(data), dimnames = NULL)
}
\arguments{
\item{dim}{Dimensions}

\item{dimnames}{Optional names for the dimensions.}
}
\value{
An array with dimensions specified by \code{dim}.
}
\description{
Creates an empty array
}
\details{
emptyArray makes a call to emptyNumeric, but is otherwise similar to \link{array} in functionality.
}
\seealso{
\link{emptyNumeric}, \link{emptyMatrix}
}
