% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emptyMatrix}
\alias{emptyMatrix}
\title{emptyMatrix}
\usage{
emptyMatrix(nrow, ncol)
}
\arguments{
\item{nrow, ncol}{number of rows and columns}
}
\value{
A numeric matrix of dim nrow x ncol.
}
\description{
Creates an uninitialized matrix (that is, assigns values according to what was allocated in the memory)
}
\examples{
mat <- emptyMatrix(10, 10)

# Initialize matrix with 1000000 entries
require(microbenchmark)
microbenchmark(matrix(, 100, 10000))
microbenchmark(emptyMatrix(100, 10000))
}
