% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{emptyNumeric}
\alias{emptyNumeric}
\title{emptyNumeric}
\usage{
emptyNumeric(n)
}
\arguments{
\item{n}{Number of elements}
}
\value{
A numeric vector of length n
}
\description{
Creates an uninitialized numeric vector (that is, assigns values according to what was allocated in the memory)
}
\details{
The calculation time is O(1) + time(memory allocation), which is handled internally by R.
For small values of n, \code{numeric} is faster, but for large values of n, \code{emptyNumeric} is much faster.
}
\examples{
vals <- numeric(1e7)
}
